Put obj_sister and obj_text_sister in rm_test
Sister is now interactable with 3 lines of text, but the box is almost
at the bottom of the screen and hard to see because its formatted for
the old room.

You can re-arrange the text location in the room by going to the draw
event for obj_text_sister and playing with the numbers.

NOTE: The dialogue and the pixel text box are drawn individually and
each have their own location. I've just been juxtaposing them on top of
each other.